+++ lib/src/IsolateArguments.dart
@@ -1,0 +1,11 @@
+import 'dart:isolate';
+
+/// Isolate scanner arguments
+class IsolateArguments {
+  SendPort sendPort;
+  String host;
+  List<int> ports;
+  Duration timeout;
+
+  IsolateArguments(this.sendPort, this.host, this.ports, this.timeout);
+}