+++ lib/tcp_scanner.dart
@@ -4,3 +4,4 @@ library tcp_scanner;
 
 export 'src/TCPScanner.dart';
 export 'src/ScanResult.dart';
+export 'src/IsolateArguments.dart';