+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 ## 1.0.1
 
 - Added `timeout` argument to TCPScanner constructor. It defines connection timeout in milliseconds while scanner waits for port response.  
+
+## 1.1.0
+- Added ability of multithread scanning
+- Added ports shuffle option
+- Fixed elapsed time calculation bug